Move Grid to own class in console app
