+++ MarsRover/MarsRover/Location.cs
@@ -16,5 +16,5 @@ namespace MarsRover
         {
             return !HasRover;
         }
-    }
+    }   
 }
